requirements.txt
jina
streamlit
pillow
torch
torchvision

